Build PDF through tinytex::latexmk() (#1222)
* use tinytex::latexmk() if available

* always use latexmk() to create PDF and never use Pandoc to create the PDF directly

this will make it faster to build PDF with keep_tex, because Pandoc does not need to convert Markdown to LaTeX twice (once for .tex, and once for .pdf)

* load quietly

* tinytex is on CRAN now, so no longer need to load it from the dark

* a news item for the PR #1222

* remove version requirement of tinytex

* always assume the output file is PDF unless it has the extension .tex

* see if _R_CHECK_TESTS_NLINES_ works (I want to seem more lines of output when the check fails)

* preinstall the LaTeX package units

* tweak comments

+++ R/render.R
@@ -647,25 +647,23 @@ render <- function(input,
       status
     }
     texfile <- file_with_ext(output_file, "tex")
-    # compile Rmd to tex when we need to generate bibliography with natbib/biblatex
-    if (need_bibtex) {
-      convert(texfile)
-      # manually compile tex if PDF output is expected
-      if (grepl('[.]pdf$', output_file)) {
+    # determine whether we need to run citeproc (based on whether we have
+    # references in the input)
+    run_citeproc <- citeproc_required(yaml_front_matter, input_lines)
+    # if the output format is LaTeX, first convert .md to .tex, and then convert
+    # .tex to .pdf via latexmk() if PDF output is requested (in rmarkdown <=
+    # v1.8, we used to call Pandoc to convert .md to .tex and .pdf separately)
+    if (output_format$pandoc$keep_tex || knitr:::is_latex_output()) {
+      # do not use pandoc-citeproc if needs to build bibliography
+      convert(texfile, run_citeproc && !need_bibtex)
+      # unless the output file has the extension .tex, we assume it is PDF
+      if (!grepl('[.]tex$', output_file)) {
         latexmk(texfile, output_format$pandoc$latex_engine, '--biblatex' %in% output_format$pandoc$args)
         file.rename(file_with_ext(texfile, "pdf"), output_file)
+        # clean up the tex file if necessary
+        if (!output_format$pandoc$keep_tex) on.exit(unlink(texfile), add = TRUE)
       }
-      # clean up the tex file if necessary
-      if ((texfile != output_file) && !output_format$pandoc$keep_tex)
-        on.exit(unlink(texfile), add = TRUE)
     } else {
-      # determine whether we need to run citeproc (based on whether we
-      # have references in the input)
-      run_citeproc <- citeproc_required(yaml_front_matter, input_lines)
-      # generate .tex if we want to keep the tex source
-      if (texfile != output_file && output_format$pandoc$keep_tex)
-        convert(texfile, run_citeproc)
-      # run the main conversion if the output file is not .tex
       convert(output_file, run_citeproc)
     }
